user exception handling is moved to the controllers level

+++ src/main/java/com/parovsky/traver/controller/UserController.java
@@ -13,6 +13,7 @@ import java.util.Map;
 
 @RestController
 public class UserController {
+
 	private final UserService userService;
 
 	@Autowired
@@ -30,31 +31,52 @@ public class UserController {
 	public ResponseEntity<UserDTO> getUser(@RequestParam Map<String, String> params) throws UserNotFoundException {
 		if (params.containsKey("email")) {
 			String email = params.get("email");
-			UserDTO userDTO = userService.getUserByEmail(email);
-			return new ResponseEntity<>(userDTO, HttpStatus.OK);
+			if (userService.isUserExistByEmail(email)) {
+				UserDTO userDTO = userService.getUserByEmail(email);
+				return new ResponseEntity<>(userDTO, HttpStatus.OK);
+			} else {
+				throw new UserNotFoundException();
+			}
 		}else if (params.containsKey("id")) {
 			Long id = Long.parseLong(params.get("id"));
-			UserDTO userDTO = userService.getUserById(id);
-			return new ResponseEntity<>(userDTO, HttpStatus.OK);
+			if (userService.isUserExist(id)) {
+				UserDTO userDTO = userService.getUserById(id);
+				return new ResponseEntity<>(userDTO, HttpStatus.OK);
+			} else {
+				throw new UserNotFoundException();
+			}
 		}
 		return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 	}
 
 	@GetMapping("/current-user")
 	public ResponseEntity<UserDTO> getCurrentUser() throws UserNotFoundException {
-		UserDTO userDTO = userService.getCurrentUserDTO();
-		return new ResponseEntity<>(userDTO, HttpStatus.OK);
+		String userEmail = userService.getCurrentUserEmail();
+		if (userService.isUserExistByEmail(userEmail)) {
+			UserDTO userDTO = userService.getUserByEmail(userEmail);
+			return new ResponseEntity<>(userDTO, HttpStatus.OK);
+		} else {
+			throw new UserNotFoundException();
+		}
 	}
 
 	@PutMapping(value = "/user", consumes = "application/json")
 	public ResponseEntity<UserDTO> updateUser(@RequestBody UserDTO userDTO) throws UserNotFoundException {
-		UserDTO user = userService.updateUser(userDTO);
-		return new ResponseEntity<>(user, HttpStatus.OK);
+		if (userService.isUserExist(userDTO.getId())){
+			UserDTO user = userService.updateUser(userDTO);
+			return new ResponseEntity<>(user, HttpStatus.OK);
+		}else {
+			throw new UserNotFoundException();
+		}
 	}
 
 	@DeleteMapping(value = "/user/{id}")
 	public ResponseEntity<Void> deleteUser(@PathVariable Long id) throws UserNotFoundException {
-		userService.deleteUser(id);
-		return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+		if (userService.isUserExist(id)) {
+			userService.deleteUser(id);
+			return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+		}else {
+			throw new UserNotFoundException();
+		}
 	}
 }
